backend/requirements.txt
firebase-admin
flask
python-dotenv
google-generativeai
reportlab==3.6.12